adds game specific wiimote ir settings.
settings are saved to and loaded from the GameConfig files if present.

they are supposed to be completely independent from the base settings and they will not show up in the base configuration dialog.

as a general comment it's curious that almost all games manage to use different calibration, apparently the wiimote sdk doesn't provide default values. the pointer will be way off if the same calibration is used for all games.

Signed-off-by: John Peterson <jpeterson57@gmail.com>