all input config settings are saved in the gameini if a game is running, the base settings are edited when a game is not running.
if a settings is not present in the gameini the base settings in the regular ini file are used instead.

it makes most sense to save input settings per game, if profiles are used it's mostly for different games, so one can see it as an auto loaded profile which doesn't hurt.

in .\Data\User\GameConfig on the git server only the important IR settings are saved, the other settings are left out of those inis.