[minor]
according to the manual st_blocks is the number of 512 byte blocks, so the size on disk should always be st_blocks*512.

this is not always the case, for example files in a mounted directory in vmware report number of 1024 byte blocks, but as there's no way to get the block size for the particular there's no workaround, so these exceptions will have to be considered bugs in stat.c without an available workaround.

source: linux.die.net/man/2/fstat