[m+]
added overall range setting for tilt and swing

added option to disable IR, in m+ games like zelda ss it's not necessary in the game and may interfere with the m+ data. it's probably not possible to attempt to sync the m+ and IR for slow movements (for fast movements it doesn't matter).

made the wiimote dialog non-modal, to tweak settings while playing it's necessary to be able to have it open without blocking the main window

fixed sensitivity setting for non-relative cursor inputs