[nogui, benchmark] benchmark flag
add nogui build config for windows.

add benchmark flag for nogui build.

dtm loads iso from gameid rather than from selected iso. ScanDirectoryTreeRecursive code moved from GameListCtrl to FileUtil for that reason.

BootManager also altered to try default and last iso if an empty iso string is provided. part of code moved from FrameTools for that reason.

(The recent fixes to Movie.cpp and State.cpp in the main clone is not merged into this branch, so it's better to test the benchmark mode with the master branch where it is merged.)